requirements.txt
biopython>=1.79
matplotlib>=3.5.0
networkx>=2.6.3
nltk>=3.6.5
numpy>=1.21.4
openpyxl>=3.0.9
pandas>=1.3.4
scikit-learn>=1.0.1
seaborn>=0.11.2
wordcloud>=1.8.1

